+++ test/Microsoft.TestPlatform.Acceptance.IntegrationTests/ProcessesInteractionTests.cs
@@ -3,6 +3,7 @@
 
 using System.IO;
 
+using Microsoft.TestPlatform.TestUtilities;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Microsoft.TestPlatform.AcceptanceTests;
