+++ test/Microsoft.TestPlatform.TestUtilities/NetCoreRunnerAttribute.cs
@@ -7,11 +7,9 @@ using System.Globalization;
 using System.Linq;
 using System.Reflection;
 
-using Microsoft.TestPlatform.TestUtilities;
-
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Microsoft.TestPlatform.AcceptanceTests;
+namespace Microsoft.TestPlatform.TestUtilities;
 
 /// <summary>
 /// Runs tests using the dotnet vstest.console.dll built against .NET Core 3.1.
